backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0
anthropic==0.8.1
python-dotenv==1.0.0
numpy==1.24.3
werkzeug==3.0.1